Some refactoring to create abstraction in front of the original highlighting functionality and the replacement functionality
